% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc_batch}
\alias{abc_batch}
\title{Single batch of ABC}
\usage{
abc_batch(sobs, priorSims, simStats, M, v = NULL)
}
\arguments{
\item{sobs}{Vector of observed summary statistics.}

\item{priorSims}{Matrix whose rows are vector of parameters drawn from the prior.}

\item{simStats}{Function mapping a vector of parameters to summary statistics.}

\item{M}{How many simulations to accept in this batch.}

\item{v}{Optional vector of estimated variances for each summary statistic}
}
\description{
Interal function carrying out part of ABC inference calculations
}
\details{
Outputs a list containing:
1) matrix whose rows are accepted parameter vectors, and a column of resulting distances
2) value of v.
If v is not supplied then variances are estimated here and returned.
(The idea is that this is done for the first batch, and these values are reused from then on.)
}

