% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\name{fx}
\alias{fx}
\title{Exchange rate example}
\usage{
fx(type = c("standard", "quick", "for paper"))
}
\arguments{
\item{type}{What type of demo to perform. \code{standard} runs the full demo displaying plots. \code{quick} is a fast demo suitable for testing the package. \code{for paper} saves pdfs which can be used in the paper.}
}
\description{
Performs a demo analysis of exchange rate data
}
\details{
\code{fx} performs the analysis of exchange rate data in the supporting reference (Prangle 2017).
}
\examples{
\dontrun{
fx()
}
}
\references{
D. Prangle gk: An R package for the g-and-k and generalised g-and-h distributions, 2017.
}
