\name{gk-package}
\alias{gk-package}
\alias{gk}
\docType{package}
\title{gk}
\description{Functions for the g-and-k distribution}
\details{
\tabular{ll}{
Package: \tab gk\cr
Type: \tab Package\cr
Version: \tab 0.35\cr
Date: \tab 2016-01-26\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Dennis Prangle

Maintainer: Dennis Prangle <dennis.prangle@gmail.com>
}
\keyword{ package }

