% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gkMLE.R
\name{gk.mle}
\alias{gk.mle}
\title{g-and-k maximum likelihood estimation}
\usage{
gk.mle(x, theta0, details = FALSE, do.message = TRUE, ...)
}
\arguments{
\item{x}{g-and-k values}

\item{theta0}{Initial guess of (A,B,g,k)}

\item{details}{If TRUE the full \code{optim} output is returned}

\item{do.message}{If TRUE a message is shown that the numerical calculations may be time consuming}

\item{...}{Further options for \code{optim}}
}
\value{
If \code{details==FALSE}, a vector of (A,B,g,k) MLEs. Otherwise, full output from \code{optim}.
}
\description{
Estimate the MLEs of A,B,g and k given independent identically distributed g-and-k values
}
\details{
The maximum likelihood estimates of (A,B,g,k) are calculated taking c=0.8. This is done by numerical optimisation of \code{dgk} which itself must be estimated numerically. Therefore this function may be time-consuming and the results guaranteed to be accurate.
}
\examples{
x <- rgk(5, A=3, B=1, g=2, k=0.5)
gk.mle(x, c(5,5,5,5))
}

