% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gk2z}
\alias{gk2z}
\title{Transform g-and-k draws to standard normal by numerical inversion.}
\usage{
gk2z(x, A, B, g, k, c = 0.8, theta = NULL)
}
\arguments{
\item{x}{A vector of g-and-k draws.}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters.  Must be positive.}

\item{g}{Vector of g parameters.}

\item{k}{Vector of k parameters.  Must be greater than -0.5.}

\item{c}{Vector of c parameters.  Often fixed at 0.8 (see Rayner and MacGillivray) which is the default.}

\item{theta}{Vector or matrix of parameter values. If this is supplied all other parameter arguments are ignored. A vector is treated as a single row matrix.  The columns may correspond to either 1) (A,B,g,k) with c taken to equal 0.8 or 2) (A,B,c,g,k).}
}
\value{
A vector of standard normal values.
}
\description{
Transform g-and-k draws to standard normal by numerical inversion.
}

