% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gk2z.scalar}
\alias{gk2z.scalar}
\title{Transform one g-and-k draw to standard normal.}
\usage{
gk2z.scalar(x, theta)
}
\arguments{
\item{x}{A g-and-k value}

\item{theta}{Vector (A,B,c,g,k).}
}
\value{
A standard normal draw.
}
\description{
Transform one g-and-k draw to standard normal.
}

