% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdsa.R
\name{logSumExp}
\alias{logSumExp}
\title{Calculate log sum exp safely}
\usage{
logSumExp(a, b)
}
\arguments{
\item{a}{Number}

\item{b}{Number}
}
\value{
Value of log(exp(a)+exp(b))
}
\description{
Calculate log(exp(a)+exp(b)) avoiding numerical issues
}
