% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{momentEstimates}
\alias{momentEstimates}
\title{Convert octiles to moment estimates}
\usage{
momentEstimates(octiles)
}
\arguments{
\item{octiles}{Vector of octiles.}
}
\value{
Vector of moment estimates.
}
\description{
Convert octiles to estimates of location, scale, skewness and kurtosis.
}
\details{
Converts octiles to robust estimate of location, scale, skewness and kurtosis as used by Drovandi and Pettitt (2011).
}
\references{
C. C. Drovandi and A. N. Pettitt. Likelihood-free Bayesian estimation of multivariate quantile distributions. Computational Statistics & Data Analysis, 2011.
}

