% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgh.R
\name{pgh_scalar}
\alias{pgh_scalar}
\title{Distribution function for the g-and-h distribution}
\usage{
pgh_scalar(q, A, B, g, h, c = 0.8, zscale = FALSE)
}
\arguments{
\item{q}{A quantiles.}

\item{A}{A (location) parameter.}

\item{B}{B (scale) parameter. Must be positive.}

\item{g}{g parameter.}

\item{h}{h parameter. Must be non-negative.}

\item{c}{c parameter. Often fixed at 0.8 which is the default.}

\item{zscale}{When true returns the N(0,1) quantile of the cdf (needed by dgh).}
}
\value{
The cumulative probability.
}
\description{
Distribution function for the g-and-h distribution
}
\details{
This internal function performs the calculation assuming scalar inputs.
 The exported function is a vectorised wrapper of this.
}
\keyword{internal}

