% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgk.R
\name{pgk_scalar}
\alias{pgk_scalar}
\title{Distribution function for the g-and-k distribution}
\usage{
pgk_scalar(q, A, B, g, k, c = 0.8, zscale = FALSE)
}
\arguments{
\item{q}{A quantiles.}

\item{A}{A (location) parameter.}

\item{B}{B (scale) parameter. Must be positive.}

\item{g}{g parameter.}

\item{k}{k parameter. Must be at least -0.5.}

\item{c}{c parameter. Often fixed at 0.8 which is the default.}

\item{zscale}{When true returns the N(0,1) quantile of the cdf (needed by dgk).}
}
\value{
The cumulative probability.
}
\description{
Distribution function for the g-and-k distribution
}
\details{
This internal function performs the calculation assuming scalar inputs.
 The exported function is a vectorised wrapper of this.
}

