% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdsa.R
\name{project}
\alias{project}
\title{Project into a region}
\usage{
project(x, xmin, xmax)
}
\arguments{
\item{x}{Vector}

\item{xmin}{Vector of lower bounds}

\item{xmax}{Vector of upper bounds}
}
\value{
Vector of closest values to x which satisfy the bounds
}
\description{
Project a vector elementwise into a constrained region
}
