% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgh.R
\name{Qgh_deriv}
\alias{Qgh_deriv}
\title{g-and-h Q derivative}
\usage{
Qgh_deriv(
  z,
  A,
  B,
  g,
  h,
  c = 0.8,
  getR = FALSE,
  type = c("generalised", "tukey")
)
}
\arguments{
\item{z}{A vector of normal quantiles.}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{h}{Vector of h parameters. Must be positive.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}

\item{getR}{When \code{TRUE} returns a faster partial calculation that has the same sign as the derivative (Used in checking parameter validity.)}

\item{type}{Can be "generalised" (default) or "tukey".}
}
\value{
The derivative of the g-and-h Q function at p.
}
\description{
Derivative of the g-and-h Q function.
}
\keyword{internal}
