% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgh.R
\name{Qgh_log_deriv}
\alias{Qgh_log_deriv}
\title{g-and-h Q log derivative}
\usage{
Qgh_log_deriv(z, A, B, g, h, c = 0.8)
}
\arguments{
\item{z}{A vector of normal quantiles.}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{h}{Vector of k parameters. Must be positive.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}
}
\value{
The derivative of the g-and-h Q function at p.
}
\description{
Derivative of the g-and-h log(Q) function.
}

