% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgk.R
\name{Qgk_deriv}
\alias{Qgk_deriv}
\title{g-and-k Q derivative}
\usage{
Qgk_deriv(z, A, B, g, k, c = 0.8, getR = FALSE)
}
\arguments{
\item{z}{A vector of normal quantiles.}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{k}{Vector of k parameters. Must be at least -0.5.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}

\item{getR}{When \code{TRUE} returns a faster partial calculation that has the same sign as the derivative (Used in checking parameter validity.)}
}
\value{
The derivative of the g-and-k Q function at p.
}
\description{
Derivative of the g-and-k Q function.
}
\keyword{internal}
