% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgk.R
\name{Qgk_log_deriv}
\alias{Qgk_log_deriv}
\title{g-and-k Q log derivative}
\usage{
Qgk_log_deriv(z, A, B, g, k, c = 0.8)
}
\arguments{
\item{z}{A vector of normal quantiles.}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{k}{Vector of k parameters. Must be greater than -0.5.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}
}
\value{
The derivative of the g-and-k Q function at p.
}
\description{
Derivative of the g-and-k log(Q) function.
}

