% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgk.orderstats}
\alias{rgk.orderstats}
\title{g&k order statistics}
\usage{
rgk.orderstats(n, orderstats, A, B, g, k, c = 0.8, theta = NULL)
}
\arguments{
\item{n}{Total number of independent draws}

\item{orderstats}{Which order statistics to generate, in increasing order}

\item{A}{Vector of A (location) parameters.}

\item{B}{Vector of B (scale) parameters.  Must be positive.}

\item{g}{Vector of g parameters.}

\item{k}{Vector of k parameters.  Must be greater than -0.5.}

\item{c}{Vector of c parameters.  Often fixed at 0.8 (see Rayner and MacGillivray) which is the default.}

\item{theta}{Vector or matrix of parameter values. If this is supplied all other parameter arguments are ignored. A vector is treated as a single row matrix.  The columns may correspond to either 1) (A,B,g,k) with c taken to equal 0.8 or 2) (A,B,c,g,k).}
}
\value{
A vector of order statistics equal in length to \code{orderstats}
}
\description{
Generates a subset of order statistics from independent g-and-k draws
}
\details{
Uniform order statistics are generaed by the exponential spacings method (see Ripley for example) and then converted to g&k values.
}
\examples{
rgk.orderstats(100, c(25,50,75), 3, 1, 2, 0.5)
}
\references{
Brian Ripley `Stochastic Simulation' Wiley (1987)
}

