% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runif.orderstats}
\alias{runif.orderstats}
\title{Uniform order statistics}
\usage{
runif.orderstats(n, orderstats)
}
\arguments{
\item{n}{Total number of independent draws}

\item{orderstats}{Which order statistics to generate, in increasing order}
}
\value{
A vector of order statistics
}
\description{
Generates a subset of order statistics from independent uniform draws
}

