% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2gh.R
\name{z2gh}
\alias{z2gh}
\title{Transform standard normal draws to g-and-h draws.}
\usage{
z2gh(z, A, B, g, h, c = 0.8)
}
\arguments{
\item{z}{Vector of N(0,1) draws.}

\item{A}{Vector of location parameters.}

\item{B}{Vector of scale parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{h}{Vector of h parameters.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}
}
\value{
A vector of g-and-h values.
}
\description{
Transform standard normal draws to g-and-h draws.
}
\keyword{internal}

