% Generated by roxygen2 (4.0.1): do not edit by hand
\name{z2gk.Jacobian}
\alias{z2gk.Jacobian}
\title{g-and-k quantile Jacobian.}
\usage{
z2gk.Jacobian(z, theta)
}
\arguments{
\item{z}{A standard normal value.}

\item{theta}{Vector (A,B,c,g,k).}
}
\value{
The derivative of the g-and-k quantile function at z.
}
\description{
Derivative of the g&k quantile function given z.
}

