% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2gk.R
\name{z2gk}
\alias{z2gk}
\title{Transform standard normal draws to g-and-k draws.}
\usage{
z2gk(z, A, B, g, k, c = 0.8)
}
\arguments{
\item{z}{Vector of N(0,1) draws.}

\item{A}{Vector of location parameters.}

\item{B}{Vector of scale parameters. Must be positive.}

\item{g}{Vector of g parameters.}

\item{k}{Vector of k parameters. Must be at least -0.5.}

\item{c}{Vector of c parameters. Often fixed at 0.8 which is the default.}
}
\value{
A vector of g-and-k values.
}
\description{
Transform standard normal draws to g-and-k draws.
}
\keyword{internal}

