% Generated by roxygen2 (4.0.1): do not edit by hand
\name{z2gk.scalar}
\alias{z2gk.scalar}
\title{Transform one normal draws to g-and-k.}
\usage{
z2gk.scalar(z, theta)
}
\arguments{
\item{z}{A standard normal draw.}

\item{theta}{Vector (A,B,c,g,k).}
}
\value{
A g-and-k value.
}
\description{
Transform one normal draws to g-and-k.
}

