% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z2gk.R
\name{z2gk.scalar}
\alias{z2gk.scalar}
\title{Transform one standard normal draw to g-and-k.}
\usage{
z2gk.scalar(z, theta)
}
\arguments{
\item{z}{A standard normal draw.}

\item{theta}{Vector (A,B,c,g,k).}
}
\value{
A g-and-k value.
}
\description{
Transform one standard normal draw to g-and-k.
}

